shiba-anime-gondola-bot/requirements.txt
beautifulsoup4
lxml
discord.py
requests